refactor: move processing functions to EodashProcess.js + remove EodashChart widget
